pyproject.toml
[project]
name = "pdf2epub"
version = "0.1.0"
description = "Convert a PDF to EPUB using Google Gemini via a simple CLI (uv project)"
authors = [
  { name = "Your Name" }
]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }
keywords = ["pdf", "epub", "gemini", "google", "cli"]

dependencies = [
  "google-generativeai>=0.7.2",
  "ebooklib>=0.18",
  "python-dotenv>=1.0",
  "rich>=13.7",
  "pdfminer.six>=20231228",
]

[project.scripts]
pdf2epub = "pdf2epub.cli:main"

[tool.uv]
package = true

[dependency-groups]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
]

[tool.ruff]
line-length = 100

[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"
